WIP move settings from `Settings` to `LocalStoreConfig`
Trying to do #5638 but this is not working very well at all.
